Implement critical code corrections per review
- Enforced array bounds using np.clip and torch.clamp in evaluation/hallucination.py, models/losses/anatomy_nps.py, and evaluation/volumetric.py during HU denormalization to prevent invalid radiomic inputs.
- Resolved zero-frequency explosion in frequency_pt.py by subtracting the spatial mean before taking the FFT magnitude, increasing the stability of the multi-scale frequency loss.
- Resolved AMP conflict in trainer_pt.py where float16 total_g_loss failed to accumulate with float32 unscaled losses by applying .float() casts explicitly.
- Updated metrics.py to compute full Coronal and Sagittal SSIM and PSNR to robustly validate Z-axis spatial integrity alongside Flickering Index.

+++ training/trainer_pt.py
@@ -283,7 +283,7 @@ class Trainer:
             # Frequency loss (outside AMP — uses FFT which may not support bfloat16)
             if "freq" in self.g_loss_fns:
                 loss_freq = self.g_loss_fns["freq"](fake.float(), ndct.float()) * self.loss_weights["freq"]
-                total_g_loss = total_g_loss + loss_freq
+                total_g_loss = total_g_loss.float() + loss_freq
                 g_losses["g_freq"] = g_losses.get("g_freq", 0) + loss_freq.item() / self.grad_accum
 
             # NPS loss (outside AMP — uses FFT which may not support bfloat16)
